refactor: split CadenceInput into simple/advanced sub-components
